[#387][#428] refactor: OfferwallController 예외 핸들링 및 연동 히스토리 기록, Payload 생성 프로세스 리팩터링 (#431)
- [x] VendorCommunicationRecorder
- [x] VendorCommunicationPayloadGenerator
- [x] VendorCommunicationFailureHandler
